reorganize bundle dict
