+++ AFQ/tests/test_api.py
@@ -707,7 +707,7 @@ def test_run_using_auto_cli():
     # after the file is written
     arg_dict['BIDS_PARAMS']['bids_path']['default'] = bids_path
     arg_dict['BIDS_PARAMS']['dmriprep']['default'] = 'vistasoft'
-    arg_dict['MAPPING']['bundle_names']['default'] = ["CST"]
+    arg_dict['DATA']['bundle_info']['default'] = ["CST"]
     arg_dict['TRACTOGRAPHY_PARAMS']['n_seeds']['default'] = 500
     arg_dict['TRACTOGRAPHY_PARAMS']['random_seeds']['default'] = True
 
@@ -827,8 +827,7 @@ def test_AFQ_data_waypoint():
             bids_path=bids_path,
             preproc_pipeline='vistasoft'),
         DATA=dict(
-            robust_tensor_fitting=True),
-        MAPPING=dict(
+            robust_tensor_fitting=True,
             bundle_info=bundle_names),
         SEGMENTATION = dict(
             scalars=[
